Move Int file to its own package for better organization and clarity
