+++ app/src/main/kotlin/com/addhen/livefront/screen/githubrepodetail/GithubRepoDetailScreen.kt
@@ -50,7 +50,7 @@ import coil.request.ImageRequest
 import com.addhen.livefront.R
 import com.addhen.livefront.data.model.GithubRepo
 import com.addhen.livefront.data.model.GithubRepo.Contributor
-import com.addhen.livefront.formatStars
+import com.addhen.livefront.extension.formatStars
 import com.addhen.livefront.screen.githubrepodetail.GithubRepoDetailViewModel.GithubRepoDetailUiState
 import com.addhen.livefront.ui.component.AppScaffold
 import com.addhen.livefront.ui.component.EmptyContent
