+++ app/src/main/kotlin/com/addhen/livefront/screen/githubrepolist/GithubRepoListScreen.kt
@@ -49,7 +49,7 @@ import coil.request.ImageRequest
 import com.addhen.livefront.connectivity.ConnectionState
 import com.addhen.livefront.R
 import com.addhen.livefront.data.model.GithubRepo
-import com.addhen.livefront.formatStars
+import com.addhen.livefront.extension.formatStars
 import com.addhen.livefront.ui.component.AppScaffold
 import com.addhen.livefront.ui.component.ConnectivityStatus
 import com.addhen.livefront.ui.component.ErrorInfo
